LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/caret-color-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSPositionValue
CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSPositionValue
